Fix embedding anchor double-append, mixed-batch pixel mapping, and varlen pooling precision
- Remove manual endoftext anchor in embed paths: Qwen3-Embedding tokenizers
  already auto-append it via add_special_tokens, so the manual append doubled
  it and shifted last-token pooling (cos 0.983 -> 0.9997 vs HF).
- Match official Qwen3-VL-Embedding format: system instruction
  "Represent the user's input." + add_generation_prompt=True.
- Pass through concatenated processor pixel_values in mixed image/text
  batches instead of per-sequence row slicing, which crashed rerank/embed.
- Handle None image entries as text-only in multimodal batches.
- Accumulate MeanPool.forward_varlen in float32 to match forward() accuracy.
- Wrap all set_context/reset_context and default device/dtype mutations in
  try/finally so global state is restored on failure.
- Add tests/test_review_fixes.py regression suite.

Known remaining issue: Qwen3-VL/Qwen2.5-VL language-model side uses plain 1D
positions instead of 3-component spatial M-RoPE (mm embedding cos ~0.954 vs
HF); M-RoPE implementation to follow.

+++ nanovllm/layers/pooler.py
@@ -148,7 +148,7 @@ class MeanPool(PoolingMethod):
     ) -> torch.Tensor:
         batch_size = cu_seqlens.shape[0] - 1
         hidden_size = hidden_states.shape[-1]
-        seq_lens = (cu_seqlens[1:] - cu_seqlens[:-1]).to(hidden_states.dtype)
+        seq_lens = (cu_seqlens[1:] - cu_seqlens[:-1]).to(torch.float32)
 
         # Segment-reduce per sequence. A global cumsum would accumulate error
         # proportional to the whole batch rather than one sequence.
@@ -156,10 +156,12 @@ class MeanPool(PoolingMethod):
             torch.arange(batch_size, device=hidden_states.device),
             (cu_seqlens[1:] - cu_seqlens[:-1]).to(torch.long),
         )
+        # Accumulate in float32 like forward() does: bfloat16 index_add_
+        # drifts by ~1e-3 relative error on long sequences.
         sums = torch.zeros(
             batch_size, hidden_size,
-            dtype=hidden_states.dtype,
+            dtype=torch.float32,
             device=hidden_states.device,
         )
-        sums.index_add_(0, segment_ids, hidden_states)
+        sums.index_add_(0, segment_ids, hidden_states.float())
         return sums / seq_lens.clamp_min(1).unsqueeze(-1)
